+++ example/ios/Share Extension/ShareViewController.swift
@@ -5,8 +5,6 @@
 //  Created by Bhagat on 25/11/22.
 //
 
-import flutter_sharing_intent
-
 // If you get no such module 'flutter_sharing_intent' error.
 // Go to Build Phases of your Runner target and
 // move `Embed Foundation Extension` to the top of `Thin Binary`.