+++ ios/Classes/SwiftFlutterSharingIntentPlugin.swift
@@ -125,86 +125,118 @@ public class SwiftFlutterSharingIntentPlugin: NSObject, FlutterStreamHandler, Fl
     }
     
     private func handleUrl(url: URL?, setInitialData: Bool) -> Bool {
-        if let url = url {
-            let appGroupId = (Bundle.main.object(forInfoDictionaryKey: kAppGroupIdKey) as? String) ?? "group.\(Bundle.main.bundleIdentifier!)"
-            let userDefaults = UserDefaults(suiteName: appGroupId)
-            let message = userDefaults?.string(forKey: kUserDefaultsMessageKey)
-            if let json = userDefaults?.object(forKey: kUserDefaultsKey) as? Data {
-                print("SwiftFlutterSharingIntentPlugin : [handleUrl] \(json)")
-            }
-            if url.fragment == "media" {
-                if let key = url.host?.components(separatedBy: "=").last,
-                   let json = userDefaults?.object(forKey: key) as? Data {
-                    let sharedArray = decode(data: json)
-                    
-                    let sharedMediaFiles: [SharingFile] = sharedArray.compactMap {
-                        guard let value = getAbsolutePath(for: $0.value) else {
-                            return nil
-                        }
-                        if ($0.type == .video && $0.thumbnail != nil) {
-                            let thumbnail = getAbsolutePath(for: $0.thumbnail!)
-                            return SharingFile.init(value: value, thumbnail: thumbnail, duration: $0.duration, type: $0.type)
-                        } else if ($0.type == .video && $0.thumbnail == nil) {
-                            return SharingFile.init(value: value, thumbnail: nil, duration: $0.duration, type: $0.type)
-                        }
-                        
-                        return SharingFile.init(value: value, thumbnail: nil, duration: $0.duration, type: $0.type)
-                    }
-                    latestSharing = sharedMediaFiles
-                    if(setInitialData) {
-                        initialSharing = latestSharing
-                    }
-                    eventSinkMedia?(toJson(data: latestSharing))
-                }
-            } else if url.fragment == "file" {
-                if let key = url.host?.components(separatedBy: "=").last,
-                   let json = userDefaults?.object(forKey: key) as? Data {
-                    let sharedArray = decode(data: json)
-                    let sharedMediaFiles: [SharingFile] = sharedArray.compactMap{
-                        guard getAbsolutePath(for: $0.value) != nil else {
-                            return nil
-                        }
-                        return SharingFile.init(value: $0.value,
-                                                thumbnail: nil, duration: nil,
-                                                type: $0.type)
-                    }
-                    latestSharing = sharedMediaFiles
-                    if(setInitialData) {
-                        initialSharing = latestSharing
-                    }
-                    eventSinkMedia?(toJson(data: latestSharing))
-                }
-            } else if url.fragment == "url" {
-                if let key = url.host?.components(separatedBy: "=").last,
-                   let sharedArray = userDefaults?.object(forKey: key) as? [String] {
-                    latestSharing = [SharingFile.init(value:  sharedArray.joined(separator: ","), thumbnail: nil, duration: nil, type:  SharingFileType.url)]
-                    if(setInitialData) {
-                        initialSharing = latestSharing
-                    }
-                    eventSinkMedia?(toJson(data: latestSharing))
-                }
-            } else if url.fragment == "text" {
-                if let key = url.host?.components(separatedBy: "=").last,
-                   let sharedArray = userDefaults?.object(forKey: key) as? [String] {
-                    latestSharing = [SharingFile.init(value:  sharedArray.joined(separator: ","), thumbnail: nil, duration: nil, type: SharingFileType.text)]
-                    if(setInitialData) {
-                        initialSharing = latestSharing
-                    }
-                    eventSinkMedia?(toJson(data: latestSharing))
-                }
-            } else {
-                latestSharing = [SharingFile.init(value: url.absoluteString, thumbnail: nil, duration: nil, type: SharingFileType.text)]
-
-                if(setInitialData) {
-                    initialSharing = latestSharing
-                }
-                eventSinkMedia?(latestSharing)
-            }
-            return true
-        }
-        latestSharing = nil
-        return false
-    }
+           let appGroupId = Bundle.main.object(forInfoDictionaryKey: kAppGroupIdKey) as? String
+           let defaultGroupId = "group.\(Bundle.main.bundleIdentifier!)"
+           let userDefaults = UserDefaults(suiteName: appGroupId ?? defaultGroupId)
+           
+           let message = userDefaults?.string(forKey: kUserDefaultsMessageKey)
+           if let json = userDefaults?.object(forKey: kUserDefaultsKey) as? Data {
+               let sharedArray = decode(data: json)
+               let sharedMediaFiles: [SharingFile] = sharedArray.compactMap {
+                   guard let value = $0.type == .text || $0.type == .url ? $0.value
+                           : getAbsolutePath(for: $0.value) else {
+                       return nil
+                   }
+                   
+                   return SharingFile(
+                    value: value,
+                       mimeType: $0.mimeType,
+                       thumbnail: getAbsolutePath(for: $0.thumbnail),
+                       duration: $0.duration,
+                       type: $0.type,
+                       message: message
+                   )
+               }
+               latestSharing = sharedMediaFiles
+               if(setInitialData) {
+                   initialSharing = latestSharing
+               }
+               eventSinkMedia?(toJson(data: latestSharing))
+           }
+           return true
+       }
+    
+//    private func handleUrl(url: URL?, setInitialData: Bool) -> Bool {
+//        if let url = url {
+//            let appGroupId = (Bundle.main.object(forInfoDictionaryKey: kAppGroupIdKey) as? String) ?? "group.\(Bundle.main.bundleIdentifier!)"
+//            let userDefaults = UserDefaults(suiteName: appGroupId)
+//            let message = userDefaults?.string(forKey: kUserDefaultsMessageKey)
+//            if let json = userDefaults?.object(forKey: kUserDefaultsKey) as? Data {
+//                print("SwiftFlutterSharingIntentPlugin : [handleUrl] \(json)")
+//            }
+//            if url.fragment == "media" {
+//                if let key = url.host?.components(separatedBy: "=").last,
+//                   let json = userDefaults?.object(forKey: key) as? Data {
+//                    let sharedArray = decode(data: json)
+//                    
+//                    let sharedMediaFiles: [SharingFile] = sharedArray.compactMap {
+//                        guard let value = getAbsolutePath(for: $0.value) else {
+//                            return nil
+//                        }
+//                        if ($0.type == .video && $0.thumbnail != nil) {
+//                            let thumbnail = getAbsolutePath(for: $0.thumbnail!)
+//                            return SharingFile.init(value: value, thumbnail: thumbnail, duration: $0.duration, type: $0.type)
+//                        } else if ($0.type == .video && $0.thumbnail == nil) {
+//                            return SharingFile.init(value: value, thumbnail: nil, duration: $0.duration, type: $0.type)
+//                        }
+//                        
+//                        return SharingFile.init(value: value, thumbnail: nil, duration: $0.duration, type: $0.type)
+//                    }
+//                    latestSharing = sharedMediaFiles
+//                    if(setInitialData) {
+//                        initialSharing = latestSharing
+//                    }
+//                    eventSinkMedia?(toJson(data: latestSharing))
+//                }
+//            } else if url.fragment == "file" {
+//                if let key = url.host?.components(separatedBy: "=").last,
+//                   let json = userDefaults?.object(forKey: key) as? Data {
+//                    let sharedArray = decode(data: json)
+//                    let sharedMediaFiles: [SharingFile] = sharedArray.compactMap{
+//                        guard getAbsolutePath(for: $0.value) != nil else {
+//                            return nil
+//                        }
+//                        return SharingFile.init(value: $0.value,
+//                                                thumbnail: nil, duration: nil,
+//                                                type: $0.type)
+//                    }
+//                    latestSharing = sharedMediaFiles
+//                    if(setInitialData) {
+//                        initialSharing = latestSharing
+//                    }
+//                    eventSinkMedia?(toJson(data: latestSharing))
+//                }
+//            } else if url.fragment == "url" {
+//                if let key = url.host?.components(separatedBy: "=").last,
+//                   let sharedArray = userDefaults?.object(forKey: key) as? [String] {
+//                    latestSharing = [SharingFile.init(value:  sharedArray.joined(separator: ","), thumbnail: nil, duration: nil, type:  SharingFileType.url)]
+//                    if(setInitialData) {
+//                        initialSharing = latestSharing
+//                    }
+//                    eventSinkMedia?(toJson(data: latestSharing))
+//                }
+//            } else if url.fragment == "text" {
+//                if let key = url.host?.components(separatedBy: "=").last,
+//                   let sharedArray = userDefaults?.object(forKey: key) as? [String] {
+//                    latestSharing = [SharingFile.init(value:  sharedArray.joined(separator: ","), thumbnail: nil, duration: nil, type: SharingFileType.text)]
+//                    if(setInitialData) {
+//                        initialSharing = latestSharing
+//                    }
+//                    eventSinkMedia?(toJson(data: latestSharing))
+//                }
+//            } else {
+//                latestSharing = [SharingFile.init(value: url.absoluteString, thumbnail: nil, duration: nil, type: SharingFileType.text)]
+//
+//                if(setInitialData) {
+//                    initialSharing = latestSharing
+//                }
+//                eventSinkMedia?(latestSharing)
+//            }
+//            return true
+//        }
+//        latestSharing = nil
+//        return false
+//    }
     
     
     public func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {